Moved Place class into new file under data
